EmailForm: moves file to src/components/contact; index: refactors import path accordingly
